test(python): move Python's Job.getChildrenValues tests to flow_tests.py
